+++ app/__init__.py
@@ -1,4 +1,4 @@
-"""BiliDownloader application package."""
+"""Bili Downloader Lite application package."""
 
-__app_name__ = "BiliDownloader"
+__app_name__ = "Bili Downloader Lite"
 __version__ = "1.0"